+++ R/get_apicarto_gpu.R
@@ -21,29 +21,29 @@
 #' allows to simplify them. See `?sf::st_simplify`
 #'
 #' @details
-#' For the moment the API cannot returned more than 5000 features.
+#' **/!\ For the moment the API cannot returned more than 5000 features.**
 #'
-#' All resssources description :
-#' * `"municipality` : information on the communes (commune with RNU, merged commune)
-#' * `"document'` : information on urban planning documents (POS, PLU, PLUi, CC, PSMV)
-#' * `"zone-urba"` : zoning of urban planning documents,
-#' * `"secteur-cc"` : communal map sectors
-#' * `"prescription-surf"` : surface prescriptions like Classified wooded area, Area contributing to the green and blue framework, Landscape element to be protected or created, Protected open space, ...
-#' * `"prescription-lin"` : linear prescription like pedestrian path, bicycle path, hedges or tree lines to be protected, ...
-#' * `"prescription-pct"` : punctual prescription like Building of architectural interest, Building to protect, Remarkable tree, Protected pools, ...
-#' * `"info-surf"` : surface information perimeters of urban planning documents like Protection of drinking water catchments, archaeological sector, noise classification, ...
-#' * `"info-lin"` : linear information perimeters of urban planning documents like Bicycle path to be created, Long hike, Façade and/or roof protected as historical monuments, ...
-#' * `"info-pct"` : punctual information perimeters of urban planning documents like Archaeological heritage, Listed or classified historical monument, Underground cavity, ...
-#' * `"acte-sup"` :
-#' * `"assiette-sup-s"` :
-#' * `"assiette-sup-l"` :
-#' * `"assiette-sup-p"` :
-#' * `"generateur-sup-s"` :
-#' * `"generateur-sup-l"` :
-#' * `"generateur-sup-p"` :
+#' All existing parameters for `ressource`  :
+#' * "municipality : information on the communes (commune with RNU, merged commune)
+#' * "document' : information on urban planning documents (POS, PLU, PLUi, CC, PSMV)
+#' * "zone-urba" : zoning of urban planning documents,
+#' * "secteur-cc" : communal map sectors
+#' * "prescription-surf" : surface prescriptions like Classified wooded area, Area contributing to the green and blue framework, Landscape element to be protected or created, Protected open space, ...
+#' * "prescription-lin" : linear prescription like pedestrian path, bicycle path, hedges or tree lines to be protected, ...
+#' * "prescription-pct" : punctual prescription like Building of architectural interest, Building to protect, Remarkable tree, Protected pools, ...
+#' * "info-surf" : surface information perimeters of urban planning documents like Protection of drinking water catchments, archaeological sector, noise classification, ...
+#' * "info-lin" : linear information perimeters of urban planning documents like Bicycle path to be created, Long hike, Façade and/or roof protected as historical monuments, ...
+#' * "info-pct" : punctual information perimeters of urban planning documents like Archaeological heritage, Listed or classified historical monument, Underground cavity, ...
+#' * "acte-sup" :
+#' * "assiette-sup-s" :
+#' * "assiette-sup-l" :
+#' * "assiette-sup-p" :
+#' * "generateur-sup-s" :
+#' * "generateur-sup-l" :
+#' * "generateur-sup-p" :
 #'
 #' @importFrom checkmate assert assert_choice check_character check_class check_null
-#' @importFrom sf read_sf st_simplify st_union st_read
+#' @importFrom sf read_sf st_simplify st_union
 #' @importFrom httr2 req_perform req_url_path_append req_url_query req_user_agent request resp_body_json resp_body_string
 #' @importFrom geojsonsf sfc_geojson geojson_sf
 #'
@@ -54,30 +54,29 @@
 #' \dontrun{
 #' library(tmap)
 #' library(sf)
-#' point <- st_sfc(st_point(c(-0.4950188466302029, 45.428039987269926)), crs = 4326)
+#' library(dplyr)
 #'
 #' # If you know the partition, all GPU features are returned, geom is override
-#' partition <- "DU_17345"
-#' poly <- get_apicarto_gpu(x = NULL, ressource = "zone-urba", partition = partition)
-#' qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)
+#' partition <- "DU_75056"
+#' zone_urba <- get_apicarto_gpu(x = NULL, ressource = "zone-urba", partition = partition)
+#' qtm(zone_urba)
 #'
-#' # If you don't know partition (only intersection between geom and GPU features is returned)
-#' poly <- get_apicarto_gpu(x = point, ressource = "zone-urba", partition = NULL)
-#' qtm(poly)+qtm(point, symbols.col = "red", symbols.size = 2)
+#' # If you don't know partition, only GPU features intersecting shape are returned
+#' point <- st_sfc(st_point(c(2.354, 48.866)), crs = 4326)
+#' zone_urba <- get_apicarto_gpu(x = point, ressource = "zone-urba", partition = NULL)
+#' qtm(zone_urba)+qtm(point)
 #'
 #' # If you wanna find partition
 #' document <- get_apicarto_gpu(point, ressource = "document", partition = NULL)
 #' partition <- unique(document$partition)
 #'
 #' # Get all prescription : /!\ prescription is different than zone-urba
-#' partition <- "DU_17345"
+#' partition <- "DU_75056"
 #' ressources <- c("prescription-surf", "prescription-lin", "prescription-pct")
 #'
-#' library(purrr)
-#' all_prescription <- map(.x = ressources,
-#'                         .f = ~ get_apicarto_gpu(point, .x, partition))
-#' }
-#'
+#' all_prescription <- lapply(ressources, get_apicarto_gpu, x = NULL, partition = partition) |>
+#'                     bind_rows()
+#'}
 get_apicarto_gpu <- function(x,
                              ressource = "zone-urba",
                              partition = NULL,
@@ -148,13 +147,13 @@ get_apicarto_gpu <- function(x,
 #' @noRd
 hit_api <- function(ressource, param){
 
-   req <- request("https://apicarto.ign.fr/api/gpu") %>%
-      req_user_agent("happign (https://paul-carteron.github.io/happign/)") %>%
-      req_url_path_append(ressource) %>%
-      req_url_query(!!!param) %>%
-      req_perform() %>%
-      resp_body_string() %>%
-      st_read()
+   req <- request("https://apicarto.ign.fr/api/gpu") |>
+      req_user_agent("happign (https://paul-carteron.github.io/happign/)") |>
+      req_url_path_append(ressource) |>
+      req_url_query(!!!param) |>
+      req_perform() |>
+      resp_body_string() |>
+      read_sf()
 
    return(req)
 }
@@ -164,9 +163,9 @@ hit_api <- function(ressource, param){
 #' @param dTolerance tolerance for simplifying
 #' @noRd
 prepare_shape <- function(x, dTolerance){
-   res <- x %>%
-      st_make_valid() %>%
-      st_union() %>%
-      st_transform(4326) %>%
+   res <- x |>
+      st_make_valid() |>
+      st_union() |>
+      st_transform(4326) |>
       st_simplify(dTolerance = dTolerance)
 }