+++ tests/testthat/test-get_wfs.R
@@ -1,109 +1,125 @@
-test_that("hit_api_wfs format bbox from sf object", {
+test_that("chek_wfs_input", {
+   expect_error(check_get_wfs_input("bad_input", "intersects"),
+                "`shape` should have class `sf`, `sfc` or `NULL`")
+   expect_error(check_get_wfs_input(point, "bad_input"),
+                "`spatial_filter` should be one of : ")
+   expect_no_error(check_get_wfs_input(point, "intersects"))
+   expect_no_error(check_get_wfs_input(point, "INTERSECTS"))
+   expect_no_error(check_get_wfs_input(point, c("intersects")))
+   expect_no_error(check_get_wfs_input(point, c("intersects", NULL)))
+   expect_no_error(check_get_wfs_input(point, c("dwithin", 50, "meters")))
+})
 
-   shape_sf <- read_sf(system.file("shape/nc.shp", package = "sf"))
-   expect_s3_class(shape_sf, "sf")
-   expect_equal(dim(shape_sf), c(100, 15))
+test_that("save_wfs", {
 
-   bbox <- st_bbox(st_transform(shape_sf, 4326))
-   expect_s3_class(bbox, "bbox")
-   expect_length(bbox, 4)
+   filename <- tempfile(fileext = ".shp")
 
-   formated_bbox <- paste(bbox["xmin"], bbox["ymin"], bbox["xmax"], bbox["ymax"],
-                          "epsg:4326",
-                          sep = ",")
-   expect_type(formated_bbox, "character")
-   expect_length(formated_bbox, 1)
-   expect_equal(length(gregexpr(",",formated_bbox, fixed = TRUE)[[1]]), 4)
+   expect_no_error(save_wfs(filename, point, F))
+   expect_error(save_wfs(filename, point, F),
+                "Dataset already exists")
+
+   warn_resp <- st_as_sf(point)
+   warn_resp$longerthant10char <- NA
+   expect_warning(save_wfs(filename, warn_resp, T),
+                  "abbreviated for ESRI Shapefile driver")
 
 })
-test_that("hit_api_wfs format bbox from sfc object", {
 
-   shape_sfc <- st_as_sfc(read_sf(system.file("shape/nc.shp", package = "sf")))
-   expect_s3_class(shape_sfc, "sfc")
-   expect_equal(length(shape_sfc), 100)
+test_that("build_wfs_req", {
+   skip_on_ci()
+   skip_on_cran()
+   all_shp_type <- lapply(list(point, multipoint, line, multiline, poly, multipoly),
+                          build_wfs_req,
+                          apikey = "altimetrie",
+                          layer_name = "ELEVATION.CONTOUR.LINE:courbe",
+                          spatial_filter = "within",
+                          crs = 4326)
+   res <- lapply(all_shp_type, function(x){x$body$data$cql_filter})
 
-   bbox <- st_bbox(st_transform(shape_sfc, 4326))
-   expect_s3_class(bbox, "bbox")
-   expect_length(bbox, 4)
+   expect_match(res[[1]], "WITHIN(the_geom, POINT (47.813 -4.34", fixed = T)
+   expect_match(res[[2]], "WITHIN(the_geom, MULTIPOINT ((47.813", fixed = T)
+   expect_match(res[[3]], "WITHIN(the_geom, LINESTRING (47.813 ", fixed = T)
+   expect_match(res[[4]], "WITHIN(the_geom, MULTILINESTRING ((4", fixed = T)
+   expect_match(res[[5]], "WITHIN(the_geom, POLYGON ((47.813 -4", fixed = T)
+   expect_match(res[[6]], "WITHIN(the_geom, MULTIPOLYGON (((47.", fixed = T)
 
-   formated_bbox <- paste(bbox["xmin"], bbox["ymin"], bbox["xmax"], bbox["ymax"],
-                          "epsg:4326",
-                          sep = ",")
-   expect_type(formated_bbox, "character")
-   expect_length(formated_bbox, 1)
-   expect_equal(length(gregexpr(",",formated_bbox, fixed = TRUE)[[1]]), 4)
+   # no shape = no spatial filter
+   req <- build_wfs_req(shape = NULL,
+                        apikey = "altimetrie",
+                        layer_name = "ELEVATION.CONTOUR.LINE:courbe",
+                        spatial_filter = "within",
+                        crs = 4326)
+   ecql_filter <- req$body$data$cql_filter
+   expect_match(ecql_filter, "")
 
+   # combine spatial and ecql filter
+   req <- build_wfs_req(shape = point,
+                        apikey = "altimetrie",
+                        layer_name = "ELEVATION.CONTOUR.LINE:courbe",
+                        spatial_filter = "within",
+                        ecql_filter = "ecql_filter1",
+                        crs = 4326)
+   ecql_filter <- req$body$data$cql_filter
+   expect_match(ecql_filter,
+                "WITHIN(the_geom, POINT (47.813 -4.344)) AND ecql_filter1",
+                fixed = T)
 })
-test_that("hit_api_wfs build request properly", {
 
-   params <- list(
-      service = "WFS",
-      version = "2.0.0",
-      request = "GetFeature",
-      outputFormat = "json",
-      srsName = "EPSG:4326",
-      typeName = "layer_name",
-      bbox = "formated_bbox",
-      startindex = "startindex",
-      count = 1000
-   )
+with_mock_dir("wfs_intersect", {
+   test_that("wfs_intersect", {
+      skip_on_cran()
+      skip_on_ci()
+      skip_if_offline()
 
-   apikey <- "VERIF"
+      apikey <- "administratif"
+      layer_name <- "LIMITES_ADMINISTRATIVES_EXPRESS.LATEST:commune"
+      spatial_filter <- "intersects"
 
-   request <- request("https://wxs.ign.fr") %>%
-      req_url_path_append(apikey) %>%
-      req_url_path_append("geoportail/wfs") %>%
-      req_user_agent("happign (https://paul-carteron.github.io/happign/)") %>%
-      req_url_query(!!!params)
+      all_shp_type <- lapply(list(point, multipoint, line, multiline, poly, multipoly),
+                             get_wfs,
+                             apikey = "administratif",
+                             layer_name = "LIMITES_ADMINISTRATIVES_EXPRESS.LATEST:commune",
+                             spatial_filter = "intersects")
 
-   expect_s3_class(request, "httr2_request")
-   expect_equal(nchar(request$url), 194)
-   expect_match(request$url, "VERIF")
+      invisible(lapply(all_shp_type,
+                       expect_s3_class, class = "sf"))
+      invisible(lapply(all_shp_type,
+                       function(x){expect_equal(dim(x), c(1,12))}))
+      invisible(lapply(all_shp_type,
+                       function(x){expect_true(st_drop_geometry(x)[1,3] == "PENMARCH")}))
+      })
+   },simplify = FALSE)
 
-})
-test_that("hit_api_wfs error", {
-   shape_sf <- read_sf(system.file("shape/nc.shp", package = "sf"))
-   layer_name <- "no_need"
-
-   expect_error(hit_api_wfs("a"))
-   expect_error(hit_api_wfs())
-   expect_error(hit_api_wfs(shape, layer_name,  1000)) # Don't forget the apikey !
-   expect_error(hit_api_wfs("parcellaire", shape, layer_name,  1000)) # Forbidden
-})
-with_mock_dir("hit_api_wfs perform request", {
-   #/!\ Again, you have to manually change encoding "UTF-8" to "ISO-8859-1" !
-   test_that("hit_api_wfs perform request", {
-      shape <- st_polygon(list(matrix(c(-4.373, -4.373, -4.372, -4.372, -4.373, 47.798,
-                                            47.799, 47.799, 47.798, 47.798), ncol = 2)))
-      shape <- st_sfc(shape, crs = st_crs(4326))
-
-      apikey <- "parcellaire"
-      layer_name <- "CADASTRALPARCELS.PARCELLAIRE_EXPRESS:parcelle"
-
-      resp <- hit_api_wfs(shape, apikey, layer_name, startindex = 0)
-      expect_s3_class(resp, "sf")
-   })
-}, simplify = FALSE)
-with_mock_dir("get_wfs simple request", {
-   #/!\ Again, you have to manually change encoding "UTF-8" to "ISO-8859-1" !
-   test_that("get_wfs", {
+with_mock_dir("wfs_ecql_filter", {
+   test_that("get_wfs_ecql", {
       skip_on_cran()
       skip_if_offline()
-      shape <- st_polygon(list(matrix(c(-4.373, -4.373, -4.372, -4.372, -4.373, 47.798,
-                                        47.799, 47.799, 47.798, 47.798), ncol = 2)))
-      shape <- st_sfc(shape, crs = st_crs(4326))
-      apikey <- "cartovecto"
-      layer_name <- "BDCARTO_BDD_WLD_WGS84G:troncon_route"
 
-      filename <- file.path(tempdir(), "troncon_route.shp")
+      apikey <- "administratif"
+      layer_name <- "LIMITES_ADMINISTRATIVES_EXPRESS.LATEST:commune"
+      spatial_filter <- NULL
+      ecql_filter <- "nom_m LIKE 'PEN%RCH' AND population < 6000"
 
-      layer <- get_wfs(shape = shape,
-                       apikey = apikey,
-                       layer_name = layer_name,
-                       file = filename)
+      resp <- get_wfs(point, apikey, layer_name, NULL, spatial_filter, ecql_filter)
+      expect_s3_class(resp, "sf")
+      expect_equal(dim(resp), c(1,12))
+      expect_true(st_drop_geometry(resp)[1,3] == "PENMARCH")
+})},
+simplify = FALSE)
 
-      expect_s3_class(layer, "sf")
-      expect_match(list.files(tempdir()), "troncon_route", fixed = TRUE, all = FALSE)
+with_mock_dir("wfs_empty", {
+   test_that("get_wfs empty_features", {
+      skip_on_cran()
+      skip_if_offline()
 
-   })
-}, simplify = FALSE)
+      apikey <- "administratif"
+      layer_name <- "LIMITES_ADMINISTRATIVES_EXPRESS.LATEST:commune"
+      ecql_filter <- "nom_m LIKE 'BADNAME'"
+
+      expect_warning(get_wfs(apikey = apikey,
+                              layer_name = layer_name,
+                              ecql_filter = ecql_filter),
+                     "No data found, NULL is returned.")
+})},
+simplify = FALSE)
+